Move to existing util class ProgressDialogs
