--- ground/src/main/java/com/google/android/ground/ui/util/AlertDialogExt.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.ground.ui.util
-
-import android.app.AlertDialog
-import androidx.fragment.app.Fragment
-import com.google.android.ground.R
-import com.google.android.ground.databinding.ProgressDialogBinding
-
-object AlertDialogExt {
-
-  fun Fragment.showProgressDialog(): AlertDialog {
-    val binding: ProgressDialogBinding = ProgressDialogBinding.inflate(layoutInflater)
-    binding.textProgressBar.text = getText(R.string.loading)
-    val dialog = AlertDialog.Builder(context).setView(binding.root).create()
-    dialog.setCanceledOnTouchOutside(false)
-    dialog.show()
-    return dialog
-  }
-}
